Fix for issue 4344. DockLayoutPanel should have been using doubles, not ints, to keep track of positions in its layout code. Review: http://gwt-code-reviews.appspot.com/132803
git-svn-id: https://google-web-toolkit.googlecode.com/svn/trunk@7388 8db76d5a-ed1c-0410-87a9-c151d255dfc7

+++ user/src/com/google/gwt/user/client/ui/DockLayoutPanel.java
@@ -378,7 +378,7 @@ public class DockLayoutPanel extends ComplexPanel implements AnimatedLayout,
   }
 
   private void doLayout() {
-    int left = 0, top = 0, right = 0, bottom = 0;
+    double left = 0, top = 0, right = 0, bottom = 0;
 
     for (Widget child : getChildren()) {
       LayoutData data = (LayoutData) child.getLayoutData();
